+++ tests/test_objective_funs.py
@@ -69,6 +69,7 @@ from desc.objectives import (
     LinearObjectiveFromUser,
     LinkingCurrentConsistency,
     MagneticWell,
+    MaximumJ,
     MeanCurvature,
     MercierStability,
     MirrorRatio,
@@ -2008,7 +2009,7 @@ class TestObjectiveFunction:
             num_quad=16,
             num_pitch=10,
         )
-        names = ["effective ripple", "Gamma_c"]
+        names = ["effective ripple", "Gamma_c", "J_s"]
         if use_bounce1d:
             names = ["old " + n for n in names]
             theta = None
@@ -2046,6 +2047,29 @@ class TestObjectiveFunction:
         np.testing.assert_allclose(
             obj.compute(eq.params_dict), grid.compress(data[names[1]])
         )
+
+        obj = MaximumJ(
+            eq,
+            grid=obj_grid,
+            nufft_eps=1e-6,
+            use_bounce1d=use_bounce1d,
+            X=X,
+            Y=Y,
+            **opts,
+        )
+
+        w0 = obj._w0
+        w1 = obj._w1
+        thresh0 = obj._thresh0
+        dJ_ds = grid.compress(data["J_s"])
+        # Shifted ReLU operation
+        dJ_ds_filtrd = (dJ_ds - thresh0) * (dJ_ds >= thresh0)
+        obj_value = w0 * jnp.sum(dJ_ds_filtrd, axis=jnp.array([1, 2])) + w1 * jnp.max(
+            dJ_ds_filtrd, axis=jnp.array([1, 2])
+        )
+
+        obj.build()
+        np.testing.assert_allclose(obj.compute(eq.params_dict), obj_value)
 
     @pytest.mark.unit
     def test_objective_compute_against_compute_ballooning(self):
@@ -3138,7 +3162,7 @@ def test_loss_function_asserts():
 
 def _reduced_resolution_objective(eq, objective, **kwargs):
     """Speed up testing suite by defining rules to reduce objective resolution."""
-    if objective in {EffectiveRipple, GammaC}:
+    if objective in {EffectiveRipple, GammaC, MaximumJ}:
         kwargs["X"] = 8
         kwargs["Y"] = 16
         kwargs["num_transit"] = 4
@@ -4023,6 +4047,17 @@ class TestObjectiveNaNGrad:
         obj = ObjectiveFunction(
             _reduced_resolution_objective(eq, GammaC, use_bounce1d=True)
         )
+        obj.build(verbose=0)
+        g = obj.grad(obj.x())
+        assert not np.any(np.isnan(g))
+
+    @pytest.mark.unit
+    def test_objective_no_nangrad_MaximumJ(self):
+        """MaximumJ."""
+        eq = get("ESTELL")
+        with pytest.warns(UserWarning, match="Reducing radial"):
+            eq.change_resolution(2, 2, 2, 4, 4, 4)
+        obj = ObjectiveFunction(_reduced_resolution_objective(eq, MaximumJ))
         obj.build(verbose=0)
         g = obj.grad(obj.x())
         assert not np.any(np.isnan(g))