+++ CHANGELOG.md
@@ -82,7 +82,10 @@ Optionally can also contract the profiles of the original ``Equilibrium`` so tha
 - Adds error for incorrect grids in ``desc.objectives.BootstrapRedlConsistency`` and when computing ``current Redl`` and ``<J*B> Redl`` compute quantities
 - Allows Redl compute quantities to use SplineProfile and updates Redl bootstrap current consistency tutorial to include a ``SplineProfile`` optimization
 - Adds automatically generated header file showing date the input file was created with `desc.vmec.VMECIO.write_vmec_input`
+- Fixes bug where ``ObjectiveFunction`` was incorrectly using ``deriv_mode="batched"`` and the heuristic-set ``jac_chunk_size`` when ``jac_chunk_size`` is given to a sub-objective, where it should have instead defaulted to ``deriv_mode="blocked"``. See #1687
+
 - Adds ``source_grid`` argument to ``desc.magnetic_fields._MagneticField.save_mgrid function`` to allow user to control the discretization of the magnetic field object being used to construct the ``mgrid`` output.
+- Adds ``desc.objectives.maxJ`` objective function which is proxy that optimized stellarators agains Trapped Electron Modes (TEM). Also adds compute function J\_s, J\_alpha, adiabatic\_J functions related to the second adiabatic invariant J\_||. Also adds a plotting function called plot\_adiabatic\_invariant.
 - Allows ``x_scale`` to be passed to ``factorize_linear_constraints`` in ``Optimizer.optimize`` through the new ``"linear_constraint_options"``.
 
 Performance Improvements